// clinic/src/main/web/lib/model/clinic.model.ts
export interface Patient {
    id?: number;
    hospitalNum?: string;
    surname?: string;
    otherNames?: string;
    statusAtRegistration?: string;
    dateRegistration?: Moment;
    dateBirth?: Moment;
    gender?: string;
    facility?: Facility;
    uuid?: string;
}

export interface OpportunisticInfection {
    id?: number;
    description?: string;
}

export interface Adhere {
    id?: number;
    description?: string;
}

export interface AdverseDrugReaction {
    id?: number;
    description?: string;
}

export interface ClinicAdhere {
    id?: number;
    adhere?: Adhere;
    uuid?: string;
}

export interface ClinicAdverseDrugReaction {
    id?: number;
    adverseDrugReaction?: AdverseDrugReaction;
    severity?: string;
    uuid?: string;
}

export interface ClinicOpportunisticInfection {
    id?: number;
    opportunisticInfection?: OpportunisticInfection;
    uuid?: string;
}

export interface Clinic {
    facility?: Facility;
    patient?: Patient;
    id?: number;
    dateVisit?: Moment;
    clinicStage?: string;
    funcStatus?: string;
    tbStatus?: string;
    viralLoad?: number;
    cd4?: number;
    cd4p?: number;
    regimenType?: any;
    regimen?: any;
    bodyWeight?: number;
    height?: number;
    waist?: number;
    bp?: string;
    archived?: boolean;
    pregnant?: boolean
    pregnancyStatus?: number;
    lmp?: Moment;
    breastfeeding?: boolean;
    oiScreened?: string;
    stiTreated?: string;
    adrScreened?: boolean;
    adherenceLevel?: string;
    commence?: boolean;
    nextAppointment?: Moment;
    notes?: string;
    bp1?: number;
    bp2?: number;
    gestationalAge?: string;
    maternalStatusArt?: string;
    extra?: any;
    uuid?: string;
}

export interface ClinicVm {
    clinic?: Clinic;
    adhereList?: Adhere[];
    adrList?: ClinicAdverseDrugReaction[];
    oiList?: OpportunisticInfection[];
}